intl: Extract YAMLTranslationIndexExtension class
