+++ src/main/resources/META-INF/plugin.xml
@@ -32,7 +32,7 @@
          on how to target different products -->
     <depends>com.intellij.modules.lang</depends>
     <depends>JavaScript</depends>
-    <depends optional="true">org.jetbrains.plugins.yaml</depends>
+    <depends optional="true" config-file="yaml.xml">org.jetbrains.plugins.yaml</depends>
     <depends optional="true" config-file="handlebars.xml">com.dmarcotte.handlebars</depends>
 
     <project-components>
@@ -57,8 +57,6 @@
         <testFinder implementation="com.emberjs.navigation.EmberTestFinder"/>
 
         <psi.referenceContributor language="JavaScript" implementation="com.emberjs.psi.EmberReferenceContributor"/>
-
-        <fileBasedIndex implementation="com.emberjs.intl.EmberTranslationIndex"/>
 
         <iconProvider implementation="com.emberjs.icons.EmberIconProvider" order="first"/>
